+++ crates/macros/src/common/field.rs
@@ -97,14 +97,8 @@ impl Field<'_> {
             env_prefix: None,
         };
 
-        if field.args.default.is_some() {
-            if field.is_nested() {
-                panic!("Cannot use defaults with `nested` configs.");
-            }
-
-            if field.is_nullable() {
-                panic!("Cannot use defaults with optional settings.");
-            }
+        if field.args.default.is_some() && field.is_nested() {
+            panic!("Cannot use defaults with `nested` configs.");
         }
 
         if field.is_required() && !field.is_nullable() {